refactor logic for internevo2hf ckpt convert
