+++ internlm/model/base_model.py
@@ -2,8 +2,14 @@ from abc import ABCMeta, abstractmethod
 
 from torch import nn
 
+from internlm.model.utils import load_src_states, merge_src_states
+
 
 class BaseModel(nn.Module, metaclass=ABCMeta):
+    """
+    Base class for all models.
+    """
+
     @staticmethod
     @abstractmethod
     def load_hf_weights(folder: str, model: nn.Module) -> None:
@@ -13,3 +19,9 @@ class BaseModel(nn.Module, metaclass=ABCMeta):
     @abstractmethod
     def convert_internevo2hf_weights(src: str, tgt: str) -> None:
         raise NotImplementedError
+
+    @staticmethod
+    def load_sharded_states(src):
+        states = merge_src_states(load_src_states(src))
+        num_shards = len(states)
+        return states, num_shards
