+++ internlm/model/modeling_internlm2.py
@@ -1,7 +1,6 @@
 # Copyright (c) InternLM. All rights reserved.
 import math
 import os
-import re
 from typing import Optional
 
 import torch
@@ -31,7 +30,6 @@ from internlm.model.utils import (
 from internlm.solver.activation_checkpoint import activation_checkpoint
 from internlm.utils.logger import get_logger
 from internlm.utils.storage_manager import get_fns, llm_load, llm_save
-from internlm.utils.utils import TensorParallelMode
 from transformers.modeling_utils import (
     SAFE_WEIGHTS_INDEX_NAME,
     SAFE_WEIGHTS_NAME,
@@ -472,7 +470,7 @@ class InternLM2(BaseModel):
         return hidden_states
 
     @staticmethod
-    def load_hf_weights(folder: str, model: nn.Module):
+    def load_hf_weights(folder: str, model: nn.Module) -> None:
         """NOTE: when loading huggingface's llama pretrained weights, you should set `adapt_hf=True` in your config."""
         assert folder is not None, "Please specify the folder of the pretrained model"
         if gpc.is_rank_for_log():
@@ -576,83 +574,7 @@ class InternLM2(BaseModel):
         internlm_accelerator.empty_cache()
 
     @staticmethod
-    def convert_internevo2hf_weights(src: str, tgt: str):
-        def _find_max_tp_pp(names):
-            ckpt_names = []
-            for name in names:
-                if name.startswith("model_t") and not name.endswith("md5"):
-                    # _t: avoid conflictint with model_config.pt
-                    ckpt_names.append(name)
-
-            max_tp, max_pp = -1, -1
-            for ckpt in ckpt_names:
-                _, tp, pp = os.path.splitext(ckpt)[0].split("_")
-                max_tp = max(max_tp, int(tp[2:]) + 1)
-                max_pp = max(max_pp, int(pp[2:]) + 1)
-
-            return max_tp, max_pp
-
-        def _find_max_wp_pp(names):
-            ckpt_names = []
-            for name in names:
-                if name.startswith("model_w") and not name.endswith("md5"):
-                    ckpt_names.append(name)
-
-            max_wp, max_pp = -1, -1
-            for ckpt in ckpt_names:
-                _, wp, pp = os.path.splitext(ckpt)[0].split("_")
-                max_wp = max(max_wp, int(wp[2:]) + 1)
-                max_pp = max(max_pp, int(pp[2:]) + 1)
-
-            return max_wp, max_pp
-
-        def load_source(src):
-            ckpt_names = get_fns(src)
-            if gpc.config.parallel.tensor.mode == TensorParallelMode.isp.name:
-                max_wp, max_pp = _find_max_wp_pp(ckpt_names)
-                # 2-d array wp_rank, pp_rank
-                states = [[None for _ in range(max_pp)] for __ in range(max_wp)]
-                for wp in tqdm(range(max_wp)):
-                    for pp in tqdm(range(max_pp)):
-                        ckpt_name = os.path.join(src, f"model_wp{wp}_pp{pp}.pt")
-                        states[wp][pp] = llm_load(ckpt_name, map_location="cpu")
-            else:
-                max_tp, max_pp = _find_max_tp_pp(ckpt_names)
-                # 2-d array tp_rank, pp_rank
-                states = [[None for _ in range(max_pp)] for __ in range(max_tp)]
-                for tp in tqdm(range(max_tp)):
-                    for pp in tqdm(range(max_pp)):
-                        ckpt_name = os.path.join(src, f"model_tp{tp}_pp{pp}.pt")
-                        states[tp][pp] = llm_load(ckpt_name, map_location="cpu")
-            return states
-
-        def merge(states):
-            merged_states = []
-            for tp_state in tqdm(states):
-                layer_shift = 0
-                shifted_state = {}
-                # shift key
-                for tp_pp_state in tp_state:
-                    _layer_shift = 0
-                    keys = list(tp_pp_state.keys())
-                    for key in keys:
-                        if key.endswith(".inv_freq"):
-                            continue
-                        match = re.search(r"\.\d+\.", key)
-                        name = key
-                        if match is not None:
-                            # layers
-                            s, e = match.span()
-                            layer_idx = int(key[s + 1 : e - 1]) + layer_shift
-                            _layer_shift = max(_layer_shift, int(key[s + 1 : e - 1]))
-                            name = key[:s] + f".{layer_idx}." + key[e:]
-                        if name.startswith("model."):
-                            name = name[6:]
-                        shifted_state[name] = tp_pp_state[key]
-                    layer_shift += _layer_shift + 1
-                merged_states.append(shifted_state)
-            return merged_states
-
+    def convert_internevo2hf_weights(src: str, tgt: str) -> None:
         def permute(qkv, num_heads, num_kv_heads, head_dim, adapt_hf=True):
             if adapt_hf:
                 return qkv
@@ -665,14 +587,12 @@ class InternLM2(BaseModel):
             qkv = rearrange(qkv, "o g n i -> o (g n i)").T
             return qkv
 
-        # load and merge states
-        states = merge(load_source(src))
-        num_shards = len(states)
-
-        # load model_config
         model_config = gpc.config.model
 
-        # convert internevo2hf state_dict
+        # load states
+        states, num_shards = InternLM2.load_sharded_states(src)
+
+        # convert state_dict
         state_dict = {}
         embedding_key_list = ["tok_embeddings.word_embeddings.weight", "tok_embeddings.weight", None]
         for layer_i in tqdm(range(model_config["num_layers"])):
@@ -704,8 +624,6 @@ class InternLM2(BaseModel):
             state_dict[f"model.layers.{layer_i}.feed_forward.w3.weight"] = torch.cat(
                 [states[i][f"layers.{layer_i}.feed_forward.w3.weight"] for i in range(num_shards)], dim=0
             )
-
-        # embedding
         if model_config["embed_split_hidden"]:
             embed_concat_dim = 1
         else:
@@ -725,10 +643,9 @@ class InternLM2(BaseModel):
             },
         )
 
-        # save state_dict to hf ckpt
+        # save state_dict to hf format
         shards, index = shard_checkpoint(state_dict, weights_name=SAFE_WEIGHTS_NAME)
         for shard_file, shard in shards.items():
             llm_save(save_path=os.path.join(tgt, shard_file), saved_obj=shard, metadata={"format": "pt"})
         if index is not None:
-            # Save the index as well
             llm_save(save_path=os.path.join(tgt, SAFE_WEIGHTS_INDEX_NAME), saved_obj=index)
