+++ internlm/model/modeling_moe.py
@@ -388,5 +388,9 @@ class Internlm1MoE(BaseModel):
         return hidden_states, moe_losses
 
     @staticmethod
-    def load_hf_weights(folder: str, model: nn.Module):
+    def load_hf_weights(folder: str, model: nn.Module) -> None:
+        raise NotImplementedError
+
+    @staticmethod
+    def convert_internevo2hf_weights(src: str, tgt: str) -> None:
         raise NotImplementedError
