+++ internlm/model/utils.py
@@ -1,6 +1,13 @@
+import os
+import re
 from typing import Any, Dict, List
 
+from tqdm import tqdm
+
+from internlm.core.context.parallel_context import global_context as gpc
 from internlm.model.modules.mha import MHA
+from internlm.utils.storage_manager import get_fns, llm_load
+from internlm.utils.utils import TensorParallelMode
 
 
 def internlm1_mha_pre_load_convert(
@@ -51,3 +58,83 @@ def convert_attn_args_to_kwargs(args, kwargs) -> Dict[str, Any]:
         kwargs["max_seqlen"] = args[3]
 
     return kwargs
+
+
+def _find_max_tp_pp(names):
+    ckpt_names = []
+    for name in names:
+        if name.startswith("model_t") and not name.endswith("md5"):
+            # _t: avoid conflictint with model_config.pt
+            ckpt_names.append(name)
+
+    max_tp, max_pp = -1, -1
+    for ckpt in ckpt_names:
+        _, tp, pp = os.path.splitext(ckpt)[0].split("_")
+        max_tp = max(max_tp, int(tp[2:]) + 1)
+        max_pp = max(max_pp, int(pp[2:]) + 1)
+
+    return max_tp, max_pp
+
+
+def _find_max_wp_pp(names):
+    ckpt_names = []
+    for name in names:
+        if name.startswith("model_w") and not name.endswith("md5"):
+            ckpt_names.append(name)
+
+    max_wp, max_pp = -1, -1
+    for ckpt in ckpt_names:
+        _, wp, pp = os.path.splitext(ckpt)[0].split("_")
+        max_wp = max(max_wp, int(wp[2:]) + 1)
+        max_pp = max(max_pp, int(pp[2:]) + 1)
+
+    return max_wp, max_pp
+
+
+def load_src_states(src):
+    ckpt_names = get_fns(src)
+    if gpc.config.parallel.tensor.mode == TensorParallelMode.isp.name:
+        max_wp, max_pp = _find_max_wp_pp(ckpt_names)
+        # 2-d array wp_rank, pp_rank
+        states = [[None for _ in range(max_pp)] for __ in range(max_wp)]
+        for wp in tqdm(range(max_wp)):
+            for pp in tqdm(range(max_pp)):
+                ckpt_name = os.path.join(src, f"model_wp{wp}_pp{pp}.pt")
+                states[wp][pp] = llm_load(ckpt_name, map_location="cpu")
+    else:
+        max_tp, max_pp = _find_max_tp_pp(ckpt_names)
+        # 2-d array tp_rank, pp_rank
+        states = [[None for _ in range(max_pp)] for __ in range(max_tp)]
+        for tp in tqdm(range(max_tp)):
+            for pp in tqdm(range(max_pp)):
+                ckpt_name = os.path.join(src, f"model_tp{tp}_pp{pp}.pt")
+                states[tp][pp] = llm_load(ckpt_name, map_location="cpu")
+    return states
+
+
+def merge_src_states(states):
+    merged_states = []
+    for tp_state in tqdm(states):
+        layer_shift = 0
+        shifted_state = {}
+        # shift key
+        for tp_pp_state in tp_state:
+            _layer_shift = 0
+            keys = list(tp_pp_state.keys())
+            for key in keys:
+                if key.endswith(".inv_freq"):
+                    continue
+                match = re.search(r"\.\d+\.", key)
+                name = key
+                if match is not None:
+                    # layers
+                    s, e = match.span()
+                    layer_idx = int(key[s + 1 : e - 1]) + layer_shift
+                    _layer_shift = max(_layer_shift, int(key[s + 1 : e - 1]))
+                    name = key[:s] + f".{layer_idx}." + key[e:]
+                if name.startswith("model."):
+                    name = name[6:]
+                shifted_state[name] = tp_pp_state[key]
+            layer_shift += _layer_shift + 1
+        merged_states.append(shifted_state)
+    return merged_states
